Adding zeolite fragments.
The zeolite fragments are stored in files cluster0*.xyz.
The first cluster that is big enough is cluster02.xyz. I have
modified this structure in cluster02_edit.xyz. From there on
this structure went through a few generations.

The atom names that are prepended with "h_" are atoms that are
supposed to be replaced with Hydrogens to terminate the structure.

The atom names that are prepended with "x_" are atoms that should
be removed from the final structure.

Ultimately a few OH groups had to be added to Si atoms to terminate
the structure.

The final structure is in cluster05.xyz. However, the positions of
atoms that terminate the structure are not optimized. In subsequent
geometry optimizations the core zeolite structure is frozen and only
the positions of the atoms in the termination layer are optimized.

Once the resulting structure is obtained the positions of the Hydrogen
atoms can be frozen to mimic the presence of the zeolite framework.