Adding "interesting" structure
The cluster09-ch-o-1.0 calculation takes many steps to complete. In
particular the gradient first seems to get small, reaching a minimum
around step 36, but then grows significantly again. This suggests that
at this point the molecular geometry goes through a significant
transition, so I want to see what that might be.