Reinterpret 2-electron integrals as interaction strengths
Note that the 2-electron integrals in the MO basis with 4
different indeces have arbitrary signs dependent on the orbital's
signs (MOs are determined up to a sign). When we calculated the
occupation numbers a lot of sign information from the wavefunction
was lost anyway, and we fix that sign based on the weak-correlation
assumption, so we suppress the sign of the integrals to ensure
the weak-correlation assumption works as intended.