Fix integral signs
Here we do the same thing as in noft_eval_2e_energy_aa.fi.
I.e. we reinterpret the integrals as an interaction strength.
In this case that doesn't actually change anything as we also
only select those integrals in which every index appears twice.
Hence the sign of the orbitals do not matter as the square of
it is alway positive.