Adding default parameters
These parameters were established by comparing the NOFT results
to Full-CI results for H2 in the cc-pvdz basis set. The Full-CI
results were obtained with the MCSCF code using an all electron
and all orbital active space.

The prefactor (c_ab) was established with H2 at a bondlength of
6 Angstrom. In this situation the maximum interaction strength
determines the correlation energy. The following energies were
obtained as a function of the prefactor:

c_ab   E_noft    E_casscf
----   -------   --------
0.0    -0.7517   -0.9986
0.45   -0.9720   -0.9986
0.50   -0.9987   -0.9986
1.0    -1.2937   -0.9986

The power (p_ab) was established at the equilibrium bondlength of H2.
That bondlength was 0.76089524 Angstrom. In this regime the occupation
numbers are close to either 1 or 0 and therefore the correlation energy
is strongly dependent on the shape of the NOFT functional near 0 and 1.
This in turn depends strongly on the power in the expression
[q(1-q)]^p_ab where q is the occupation number. The following energies
were obtained as a function of the power with the prefactor set to 0.5:

p_ab     E_noft        E_casscf
-----    -----------   ---------------
0.5      -1.16832977   -1.163672981196
0.515    -1.16383554   -1.163672981196
0.5155   -1.16368726   -1.163672981196
0.5156   -1.16365765   -1.163672981196
0.5157   -1.16362806   -1.163672981196
0.516    -1.16353936   -1.163672981196
0.52     -1.16235534   -1.163672981196
0.56     -1.15203106   -1.163672981196
1.0      -1.1286       -1.163672981196

based on which p_ab was chosen to be 0.5155.