Adding the no-optimize option to the string method
Sometimes you want to optimize a reaction pathway with the string method
using one energy expression, and then evaluate a second energy expression
for the points on the same path. This option allows you to turn the
path optimization off. Instead the code runs just the energy evaluation
for all beads.